fix: add static_cast for remaining Qt6 narrowing conversions
- Fix narrowing conversions in mainwindow.cpp (3 locations)
- Fix narrowing conversions in appwindow.cpp (2 locations)
- All related to qsizetype/long long to int conversions

+++ src/appwindow.cpp
@@ -447,7 +447,7 @@ void AppWindow::openTab(const MainWindow::EditorStatus &status, bool duplicate, 
 void AppWindow::openTabs(const QStringList &paths)
 {
     LOG_INFO("Paths " << paths.join(" "));
-    int length = paths.length();
+    int length = static_cast<int>(paths.length());
 
     QProgressDialog progress(this);
     progress.setWindowModality(Qt::WindowModal);
@@ -955,7 +955,7 @@ void AppWindow::onEditorFileChanged()
                     }
                 }
             }
-            int removeLength = longestCommonPrefix.lastIndexOf('/') + 1;
+            int removeLength = static_cast<int>(longestCommonPrefix.lastIndexOf('/') + 1);
             for (auto index : tabs)
             {
                 ui->tabWidget->setTabText(index, windowAt(index)->getTabTitle(tabs.size() > 1, true, removeLength));
